requirements.txt
# Core
sentence-transformers>=2.7.0
transformers>=4.41.0,<5.0.0
torch>=2.1.0
numpy>=1.24
scikit-learn>=1.3

# App
gradio>=4.44.0,<4.46
PyYAML>=6.0

# Dev & Tests
pytest>=8.2